Chapter 17

+++ 15-10/15-10.cpp
@@ -30,7 +30,7 @@ int sumDigits(int i)
 		exp++;
 	}
 
-	int tens = pow(10, exp);
+	int tens = (int)pow(10, exp);
 	int remainder = i - (base * tens);
 	int firstDigit = i / tens;
